refactor: organize controllers and routes by domain
