+++ MJ_FB_Backend/src/app.ts
@@ -8,22 +8,22 @@ import bookingsRoutes from './routes/bookings';
 import holidaysRoutes from './routes/holidays';
 import blockedSlotsRoutes from './routes/blockedSlots';
 import breaksRoutes from './routes/breaks';
-import staffRoutes from './routes/staff';
-import adminStaffRoutes from './routes/adminStaff';
-import volunteerRolesRoutes from './routes/volunteerRoles';
-import volunteersRoutes from './routes/volunteers';
-import volunteerBookingsRoutes from './routes/volunteerBookings';
-import volunteerMasterRolesRoutes from './routes/volunteerMasterRoles';
+import staffRoutes from './routes/admin/staff';
+import adminStaffRoutes from './routes/admin/adminStaff';
+import volunteerRolesRoutes from './routes/volunteer/volunteerRoles';
+import volunteersRoutes from './routes/volunteer/volunteers';
+import volunteerBookingsRoutes from './routes/volunteer/volunteerBookings';
+import volunteerMasterRolesRoutes from './routes/volunteer/volunteerMasterRoles';
 import authRoutes from './routes/auth';
 import rolesRoutes from './routes/roles';
 import donorsRoutes from './routes/donors';
-import donationsRoutes from './routes/donations';
+import donationsRoutes from './routes/warehouse/donations';
 import clientVisitsRoutes from './routes/clientVisits';
-import surplusRoutes from './routes/surplus';
-import pigPoundsRoutes from './routes/pigPounds';
-import outgoingReceiversRoutes from './routes/outgoingReceivers';
-import outgoingDonationsRoutes from './routes/outgoingDonations';
-import warehouseOverallRoutes from './routes/warehouseOverall';
+import surplusRoutes from './routes/warehouse/surplus';
+import pigPoundsRoutes from './routes/warehouse/pigPounds';
+import outgoingReceiversRoutes from './routes/warehouse/outgoingReceivers';
+import outgoingDonationsRoutes from './routes/warehouse/outgoingDonations';
+import warehouseOverallRoutes from './routes/warehouse/warehouseOverall';
 import eventsRoutes from './routes/events';
 import { initializeSlots } from './data';
 import logger from './utils/logger';
