+++ MJ_FB_Backend/src/routes/admin/adminStaff.ts
@@ -6,10 +6,10 @@ import {
   updateStaff,
   deleteStaff,
   searchStaff,
-} from '../controllers/adminStaffController';
-import { authMiddleware, authorizeRoles, authorizeAccess } from '../middleware/authMiddleware';
-import { validate } from '../middleware/validate';
-import { createStaffSchema, updateStaffSchema } from '../schemas/staffSchemas';
+} from '../../controllers/admin/adminStaffController';
+import { authMiddleware, authorizeRoles, authorizeAccess } from '../../middleware/authMiddleware';
+import { validate } from '../../middleware/validate';
+import { createStaffSchema, updateStaffSchema } from '../../schemas/admin/staffSchemas';
 
 const router = Router();
