+++ MJ_FB_Backend/src/routes/volunteer/volunteerMasterRoles.ts
@@ -1,0 +1,9 @@
+import { Router } from 'express';
+import { listVolunteerMasterRoles } from '../../controllers/volunteer/volunteerMasterRoleController';
+import { authMiddleware, authorizeRoles } from '../../middleware/authMiddleware';
+
+const router = Router();
+
+router.get('/', authMiddleware, authorizeRoles('staff'), listVolunteerMasterRoles);
+
+export default router;
